tweet/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
cssselect2==0.6.0
Django==4.0.4
easy-thumbnails==2.8.1
lxml==4.9.0
Pillow==9.1.1
reportlab==3.6.9
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
tzdata==2022.1
webencodings==0.5.1
